Added integration between user registration and database
The integration is fucked up. Note, if you set fields as hidden in the datasource, the widget displaying that datasource will
have no information whatsoever that those fields exist. Thus, if you want to have hidden fields, but you need the widget to set their valu$

Secondly, smartGWT form will execute updates if you try to submit it when any ds primary key is set. So, in essence, you do not have contr$